chore: migrate NewIndexedCursor to IndexedCursor
