+++ libcobj/app/src/main/java/jp/osscons/opensourcecobol/libcobj/file/CobolIndexedFile.java
@@ -32,7 +32,7 @@ import org.sqlite.SQLiteErrorCode;
 
 /** TODO: 準備中 */
 public class CobolIndexedFile extends CobolFile {
-    private Optional<NewIndexedCursor> cursor;
+    private Optional<IndexedCursor> cursor;
     private boolean updateWhileReading = false;
     private boolean indexedFirstRead = true;
     private boolean callStart = false;
@@ -566,12 +566,12 @@ public class CobolIndexedFile extends CobolFile {
         boolean isDuplicate = this.keys[p.key_index].getFlag() != 0;
 
         this.cursor =
-                NewIndexedCursor.createCursor(p.connection, p.key, p.key_index, isDuplicate, cond);
+                IndexedCursor.createCursor(p.connection, p.key, p.key_index, isDuplicate, cond);
         if (!this.cursor.isPresent()) {
             return COB_STATUS_30_PERMANENT_ERROR;
         }
 
-        NewIndexedCursor cursor = this.cursor.get();
+        IndexedCursor cursor = this.cursor.get();
         Optional<FetchResult> optionalResult = cursor.next();
         if (optionalResult.isPresent()) {
             FetchResult result = optionalResult.get();
@@ -735,7 +735,7 @@ public class CobolIndexedFile extends CobolFile {
         boolean isDuplicate = this.keys[p.key_index].getFlag() != 0;
         if (this.indexedFirstRead || this.flag_begin_of_file) {
             this.cursor =
-                    NewIndexedCursor.createCursor(
+                    IndexedCursor.createCursor(
                             p.connection, p.key, p.key_index, isDuplicate, COB_GE);
             if (!this.cursor.isPresent()) {
                 return COB_STATUS_10_END_OF_FILE;
@@ -743,7 +743,7 @@ public class CobolIndexedFile extends CobolFile {
             this.cursor.get().moveToFirst();
         } else if (this.flag_end_of_file) {
             this.cursor =
-                    NewIndexedCursor.createCursor(
+                    IndexedCursor.createCursor(
                             p.connection, p.key, p.key_index, isDuplicate, COB_LE);
             if (!this.cursor.isPresent()) {
                 return COB_STATUS_30_PERMANENT_ERROR;
@@ -754,8 +754,8 @@ public class CobolIndexedFile extends CobolFile {
             if (!this.cursor.isPresent()) {
                 return COB_STATUS_30_PERMANENT_ERROR;
             }
-            NewIndexedCursor oldCursor = this.cursor.get();
-            Optional<NewIndexedCursor> newCursor = oldCursor.reloadCursor();
+            IndexedCursor oldCursor = this.cursor.get();
+            Optional<IndexedCursor> newCursor = oldCursor.reloadCursor();
             if (!newCursor.isPresent()) {
                 this.cursor = Optional.of(oldCursor);
             } else {
@@ -768,7 +768,7 @@ public class CobolIndexedFile extends CobolFile {
             return COB_STATUS_30_PERMANENT_ERROR;
         }
 
-        NewIndexedCursor cursor = this.cursor.get();
+        IndexedCursor cursor = this.cursor.get();
 
         final CursorReadOption cursorOpt;
         if ((readOpts & COB_READ_PREVIOUS) != 0) {
@@ -1051,7 +1051,7 @@ public class CobolIndexedFile extends CobolFile {
         p.write_cursor_open = true;
 
         if (this.access_mode == COB_ACCESS_SEQUENTIAL
-                && !NewIndexedCursor.matchKeyHead(p.key, DBT_SET(this.keys[0].getField()))) {
+                && !IndexedCursor.matchKeyHead(p.key, DBT_SET(this.keys[0].getField()))) {
             return COB_STATUS_21_KEY_INVALID;
         }
